+++ .dockerignore
@@ -1,0 +1,8 @@
+.git
+.venv
+__pycache__
+parser/.venv
+parser/__pycache__
+*.pyc
+*.exe
+parser/instrumented_*
